requirements.txt
Quart
discord.py
discord.ui
jishaku
asyncio
aiohttp
datetime
flask
typing
psutil
collection
pathlib
pymongo
wavelink
httpx
discord.ext-context
tasksio
motor
Augmentor
numpy
aiosqlite
gtts
langdetect
python-dotenv
pyttsx3
PyYAML
requests
openai
duckduckgo-search==6.3.6
pillow==11.0.0
colorama
chess
python-dateutil
topggpy
google-generativeai 
groq